project/requirements.txt
pydantic
requests
google-generativeai
streamlit
